+++ ow_lander/src/ow_lander/trajectory_planner.py
@@ -20,7 +20,7 @@ from moveit_msgs.srv import GetPositionFK
 
 from ow_lander import constants
 from ow_lander.common import Singleton, is_shou_yaw_goal_in_range
-from ow_lander.transforms import FrameTransformer
+from ow_lander.frame_transformer import FrameTransformer
 
 def _cascade_plans(plan1, plan2):
     """Joins two robot motion plans into one
